Move IoC and Views
